Initial Push (minus tests)
This is a working simple XOR Encryption program. It is a copy of the old
version that I had made in C a long time ago. I decided to begin the
project again in C# and give it a nice GUI as well as multiple file
support and multiple encryption support.